requirements-render.txt
# Core deployment requirements
cffi>=1.15.1
wheel>=0.37.1
setuptools>=65.5.1

# Flask web framework
Flask
flask-cors
flask-sqlalchemy
Flask-Login>=0.6.3
Flask-WTF
markupsafe
itsdangerous
click
Jinja2

# Cryptography for file encryption/decryption
cryptography

# Requests for HTTP file download
requests

# Watchdog for real-time folder monitoring
watchdog

# psutil for process management
psutil

# Data analysis
pandas
matplotlib
seaborn
numpy<1.26.0
scipy<1.13.0
scikit-learn>=1.3.0

# Optional: clamd for ClamAV integration on Linux/macOS
clamd>=1.0

# For automated testing
pytest>=7.0

# For loading .env files
python-dotenv>=1.0

# These Windows-specific packages won't be used on Render (Linux) but included for completeness
# pywin32 and wmi are Windows-only and will be skipped on Linux
pywin32-ctypes>=0.2.0

# Werkzeug for secure_filename
Werkzeug>=2.0

# Archive extraction support
rarfile>=4.0

# YARA for signature scanning
yara-python>=4.0

dnslib>=0.9.26

# For secure password hashing
bcrypt>=4.0

# For TOTP-based 2FA
pyotp>=2.9
py7zr>=0.21.1

# For DNS-based reputation services
dnspython>=2.0

# Machine Learning dependencies
joblib<1.3.0
requests>=2.31.0
psutil>=5.9.0

# Network monitoring and analysis
flask-limiter<4.0
limits<6.0
ordered-set<4.2
rich<14.0
pygments<3.0
typing_extensions<4.14
wrapt<1.18
markdown-it-py<3.1
mdurl<0.3
yara-python
python-magic>=0.4.27
pefile>=2022.5.30

# For VirusTotal API integration
# (requests already included)

# For inspecting file metadata
pefile>=2023.2.7
redis>=5.0.1
